Turn API protocols into concrete types
With delegate implementations. This allows for having properties of
Window type, for instance, and generally using a concrete type is easier
for the user. It also provides a place to extract non-implementation
specific functionality at some point in the future.

This allows us to compile with a Property<Window> type, but the
implementation doesn't quite work yet.