Keep reference to State in Application
Internally, this is so a Window can access the list of Screens to determine
which one it is on, for example. The reference is kept weakly on the
delegates to prevent reference cycles.

Externally, it's more of a convenience and is a strong reference, so
holding any Application or Window keeps the whole State alive.

This neccessitates changing most of ApplicationSpec to using
OSXApplicationDelegate directly, rather than through Application.