Retry applications that fail to watch up to 3 times
On my system there would be some apps that time out while trying to
watch them. This seems to fix that issue for apps that actually do
respond eventually.

A better solution would be to make this configurable, and a solution
that's better still would be to use exponential backoff.